requirements.txt
Django
djangorestframework
celery
redis
django-cors-headers
djangorestframework-simplejwt
python-decouple
pandas
openpyxl
celery[redis]
